requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# File processing
zipfile-deflate64>=0.2.0
jsonschema>=4.17.0

# Visual processing and screenshots
playwright>=1.40.0
selenium>=4.15.0
html2image>=2.0.0
wkhtmltopdf>=0.2

# Report generation
jinja2>=3.1.0
weasyprint>=60.0
reportlab>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# PDF generation
fpdf2>=2.7.0
pdfkit>=1.0.0

# HTML/CSS processing
beautifulsoup4>=4.12.0
cssutils>=2.7.0

# Data validation
pydantic>=2.0.0
marshmallow>=3.20.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional dependencies for advanced features
# AI/ML capabilities
scikit-learn>=1.3.0
opencv-python>=4.8.0  # For computer vision analysis

# Web3 analysis (future-proofing)
web3>=6.11.0

# Advanced visualization
plotly>=5.17.0
dash>=2.14.0  # For interactive dashboards